[agant ] interdit de replacer l'image la creation d'un fichier BOOL_DISABLE_IMG interdit dans tout les cas de reinstaller l'image de synchronisation avec la base. interdit toute mise a jour de l'agent. attention cela ne conserne pas les plugins.

+++ pulse_xmpp_agent/agentxmpp.py
@@ -3327,6 +3327,9 @@ class MUCBot(ClientXMPP):
         return disk + "\\\\".join(result)
 
     def reinstall_agent(self):
+        BOOL_DISABLE_IMG = os.path.join(self.pathagent, "BOOL_DISABLE_IMG")
+        if os.path.exists(BOOL_DISABLE_IMG):
+            return
         file_put_contents(
             os.path.join(self.pathagent, "BOOL_UPDATE_AGENT"),
             "use file boolean update. enable verify update.",
